Added sortDropdownBuilder to habits.js
