+++ html-pages/habits.html
@@ -52,8 +52,7 @@
           <div class="filter-group">
             <h3 class="filter-title">Sortera</h3>
             <label for="sortering"></label>
-            <select class="dropdown">
-              <option value="">Välj sortering</option>
+            <select id="sort-dropdown" class="dropdown">
             </select>
           </div>
